build(windows): compile-time scaffolding for native Windows support (#39 stage 1)
Stage 1 of #39: make the codebase compile cleanly for GOOS=windows and
add a mandatory Windows CI build job. The agent, shim, and `jitenv run`
paths return "not yet implemented" errors at runtime; real Windows
support (named-pipe transport, token-based peer check, PowerShell hook,
%LOCALAPPDATA% file modes) is Stage 2+.

Files split with `//go:build` constraints:
- internal/agent: peer_windows.go (new), socket_unix.go/socket_windows.go
  (new, listener-bind extracted from server.go), daemonize_unix.go/
  daemonize_windows.go (new, SpawnDaemon split out of daemonize.go),
  paths_unix.go/paths_windows.go (new, runtimeBaseDir helper extracted
  from paths.go).
- internal/run: run_unix.go/run_windows.go (new, replaceProcess split
  out of run.go).
- internal/shim: shim_unix.go/shim_windows.go (new, execReal split out
  of shim.go).
- internal/cli/unlock.go gets an early `runtime.GOOS == "windows"`
  guard so the user sees a sensible message instead of raw stub errors.
- e2e/cmd/unlock/main.go marked `!windows` (e2e helper, Linux-only by
  design).
- Tests that need Unix sockets, fork+exec, symlinks, or bash/zsh get
  `//go:build !windows` constraints so the Windows test job compiles
  the cross-platform surface (config, crypto, version, sources, tui,
  runnotice, agentwarn, agent's pure-Go bits) without trying to run
  the e2e suites.

CI: new `test-windows` job on `windows-latest` running `go build`,
`go vet`, and `go test ./...`. Mandatory (not allow-failure) so a
regression in the cross-platform compile fails immediately.

Not in scope (deferred to Stage 2+): named pipes, token-based peer
auth, PowerShell hook installer, Windows addition to the goreleaser
build matrix.
