+++ internal/chpwd/chpwd_test.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package chpwd
 
 import (
