+++ internal/run/run_e2e_test.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package run_test
 
 import (
