+++ internal/shell/hook_e2e_test.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package shell_test
 
 import (
