+++ internal/shell/hook_unlock_lock_test.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package shell_test
 
 import (
