+++ internal/shell/install_test.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package shell
 
 import (
