+++ internal/shim/shim_test.go
@@ -1,3 +1,5 @@
+//go:build !windows
+
 package shim_test
 
 import (
